+++ R/block_weights_heatmap.R
@@ -8,13 +8,15 @@
 #'
 #' @examples
 #' data(NCI60)
-#' mcia_results <- nipals_multiblock(data_blocks, metadata = metadata_NCI60,
-#' num_PCs = 10, plots = "none", tol = 1e-12)
+#' data_blocks_mae <- simple_mae(data_blocks,row_format="sample",
+#'                               colData=metadata_NCI60)
+#' mcia_results <- nipals_multiblock(data_blocks_mae, num_PCs = 10,
+#'                                   plots = "none", tol = 1e-12)
 #' block_weights_heatmap(mcia_results)
 #' @return heatmap object containing the block weights as a heatmap
 #' @export
 block_weights_heatmap <- function(mcia_results) {
-    bs_weights <- as.matrix(data.frame(mcia_results$block_score_weights))
+    bs_weights <- as.matrix(data.frame(mcia_results@block_score_weights))
     colnames(bs_weights) <- seq_len(ncol(bs_weights))
 
     ComplexHeatmap::Heatmap(bs_weights, cluster_columns = FALSE,