+++ R/global_scores_eigenvalues_plot.R
@@ -8,17 +8,19 @@
 #'
 #' @examples
 #' data(NCI60)
-#' mcia_results <- nipals_multiblock(data_blocks, metadata = metadata_NCI60,
-#'                                   num_PCs = 10, plots = "none", tol=1e-12)
+#' data_blocks_mae <- simple_mae(data_blocks,row_format="sample",
+#'                              colData=metadata_NCI60)
+#' mcia_results <- nipals_multiblock(data_blocks_mae, num_PCs = 10,
+#'                                  plots = "none", tol=1e-12)
 #' global_scores_eigenvalues_plot(mcia_results)
 #' @return Displays the contribution plot using eigenvalues
 #' @importFrom graphics barplot
 #' @export
 global_scores_eigenvalues_plot <- function(mcia_results) {
     # getting total variance if supplied
-    if (is.list(mcia_results$block_variances)) {
+    if (is.list(mcia_results@block_variances)) {
         ylabel <- "Prop. Total Variance"
-        totvar <- sum(unlist(mcia_results$block_variances))
+        totvar <- sum(unlist(mcia_results@block_variances))
 
     } else {
         ylabel <- "Eigenvalue"
@@ -26,8 +28,8 @@ global_scores_eigenvalues_plot <- function(mcia_results) {
     }
 
     # extract eigenvalues
-    barploteigs <- (unlist(mcia_results$eigvals)) / totvar
-    names(barploteigs) <- seq(1, length(mcia_results$eigvals))
+    barploteigs <- (unlist(mcia_results@eigvals)) / totvar
+    names(barploteigs) <- seq(1, length(mcia_results@eigvals))
 
     # generate barplot
     barplot(barploteigs, xlab = "Factors", ylab = ylabel, cex.names = 1,