+++ R/global_scores_heatmap.R
@@ -13,27 +13,27 @@ global_scores_heatmap <- function(mcia_results, color_col = NULL,
                                   color_pal = scales::viridis_pal,
                                   color_pal_params = list(option = "D")) {
     # extract the global scores
-    global_scores <- mcia_results$global_scores
+    global_scores <- mcia_results@global_scores
 
     # extract and re-label global scores
     colnames(global_scores) <- paste0("F", seq(1, ncol(global_scores)))
 
     if (!is.null(color_col)) {
         # check if metadata is within the MCIA results
-        if (!"metadata" %in% names(mcia_results)) {
-            msg <- paste("Cannot produce a sample color_col without metadata.",
-                         "Please re-run global_scores_heatmap after setting",
-                         "this information.")
-            stop(msg)
+        if (nrow(mcia_results@metadata) == 0) {
+          msg <- paste("Cannot produce a sample color_col without metadata.",
+                       "Please re-run global_scores_heatmap after setting",
+                       "this information.")
+          stop(msg)
         }
 
         # extract the column names (if not provided via color_col)
         if (typeof(color_col) == "double") {
-            color_col <- names(mcia_results$metadata)[color_col]
+            color_col <- names(mcia_results@metadata)[color_col]
         }
 
         # extract color type and create a color palette list
-        cat_values <- mcia_results$metadata[, color_col]
+        cat_values <- mcia_results@metadata[, color_col]
 
         cat_colors <- get_metadata_colors(mcia_results,
                                           color_col = color_col,