+++ R/vis_load_plot.R
@@ -10,9 +10,11 @@
 #' @return Plot of MCIA feature loadings for chosen axes
 #' @examples
 #' data(NCI60)
-#' mcia_results <- nipals_multiblock(data_blocks, metadata = metadata_NCI60,
-#' num_PCs = 10, plots = "none", tol = 1e-12)
-#' colors_omics <- get_colors(mcia_results)
+#' data_blocks_mae <- simple_mae(data_blocks,row_format="sample",
+#'                               colData=metadata_NCI60)
+#' mcia_results <- nipals_multiblock(data_blocks_mae, num_PCs = 10,
+#'                                   plots = "none", tol = 1e-12)
+#'                                   colors_omics <- get_colors(mcia_results)
 #' vis_load_plot(mcia_results, axes = c(1, 4), colors_omics = colors_omics)
 #' @importFrom ggplot2 ggplot aes_string geom_point theme_bw scale_color_manual
 #' @importFrom ggplot2 xlab ylab labs
@@ -20,11 +22,11 @@
 
 vis_load_plot <- function(mcia_out, axes = c(1, 2), colors_omics) {
     # extracting global loadings
-    gl <- mcia_out$global_loadings
-    
-    omic_dims <- vapply(mcia_out$block_loadings, dim, numeric(2))[1, ]
+    gl <- mcia_out@global_loadings
+
+    omic_dims <- vapply(mcia_out@block_loadings, dim, numeric(2))[1, ]
     omic_type <- c()
-    omics_labels <- names(mcia_out$block_loadings)
+    omics_labels <- names(mcia_out@block_loadings)
     for (i in seq_along(omics_labels)) {
         omic_label <- omics_labels[i]
         length_omic <- omic_dims[i]
@@ -38,15 +40,12 @@ vis_load_plot <- function(mcia_out, axes = c(1, 2), colors_omics) {
                         paste0("Axis_", axes[2]), "omic")
 
     # plot data
-    omic<-gl_f$omic
+    omic <- gl_f$omic
     ggplot(data = gl_f,
-            aes(x = gl_f[[colnames(gl_f)[1]]],
-                y = gl_f[[colnames(gl_f)[2]]],
-                color = omic)) +
-            geom_point(alpha = 0.3) +
-            labs(x = paste0("Axis ", axes[1]),
-                 y = paste0("Axis ", axes[2])) +
-            scale_color_manual(values = colors_omics) +
-            theme_bw()
-
+           aes(x = gl_f[[colnames(gl_f)[1]]], y = gl_f[[colnames(gl_f)[2]]],
+               color = omic)) +
+      geom_point(alpha = 0.3) +
+      labs(x = paste0("Axis ", axes[1]), y = paste0("Axis ", axes[2])) +
+      scale_color_manual(values = colors_omics) +
+      theme_bw()
 }