+++ tests/testthat/test-global_scores_heatmap.R
@@ -20,7 +20,7 @@ test_that("color column not in metadata", {
 test_that("checking the ggplot object with inputs", {
     # capture the plotting function
     p <- global_scores_heatmap(mcia_results, color_col = "cancerType")
-    expect_equal(matrix(p@matrix), matrix(mcia_results$global_scores))
+    expect_equal(matrix(p@matrix), matrix(mcia_results@global_scores))
     expect_equal("cancerType", p@right_annotation@anno_list$ColorType@label)
 })