+++ tests/testthat/test-predict_gs.R
@@ -1,5 +1,5 @@
-preds <- predict_gs(mcia_results, data_blocks)
+preds <- predict_gs(mcia_results, data_blocks_mae)
 
 test_that("correct dimensions", {
-  expect_equal(nrow(preds), nrow(mcia_results$metadata))
+  expect_equal(nrow(preds), nrow(mcia_results@metadata))
 })